[INLONG-5739][Manager] Add java doc and rename class name
